Fix 14.10.2019

+++ src/main/java/com/congdat/notaryweb/api/RoleApi.java
@@ -10,28 +10,41 @@ import java.util.List;
 
 @RestController
 @RequestMapping(value = "/api/v1/role")
-public class RoleApi {
+public class RoleApi extends AbstractApi<Role, Long, String> {
 
 		@Autowired
 		private RoleService roleService;
 
+		@Override
 		@GetMapping
 		public ResponseEntity findAll() {
 				return ResponseEntity.ok(roleService.findAll());
 		}
 
+		@Override
 		@PostMapping
-		ResponseEntity save(@RequestBody Role role) {
-				return ResponseEntity.ok(roleService.save(role));
+		public ResponseEntity save(@RequestBody Role role, @RequestParam(name = "username") String username) {
+				return ResponseEntity.ok(roleService.save(role, username));
 		}
 
+		@Override
 		@PutMapping(value = "/{id}")
-		ResponseEntity update(@RequestBody Role role, @PathVariable(name = "id") Long id) {
-				return ResponseEntity.ok(roleService.update(role, id));
+		public ResponseEntity update(@RequestBody Role role,
+																															@PathVariable(name = "id") Long id,
+																															@RequestParam(name = "username") String username) {
+				if (roleService.update(role, id, username) == null) {
+						ResponseEntity.badRequest().body("Can't find role by id: " + id);
+				}
+				return ResponseEntity.ok(roleService.update(role, id, username));
 		}
 
-		@DeleteMapping
-		public void delete(@RequestParam(name = "id") Long id) {
+		@Override
+		@DeleteMapping(value = "/{id}")
+		public ResponseEntity delete(@PathVariable(name = "id") Long id) {
+				if (!roleService.delete(id)) {
+						return ResponseEntity.badRequest().body("Can't find role by id: " + id);
+				}
 				roleService.delete(id);
+				return ResponseEntity.ok().body("Role id: " + id + " was deleted");
 		}
 }
